refactor: Update utilities import in conformer_generation scripts
