+++ scripts/pocket_finding/main.py
@@ -11,10 +11,11 @@ from scripts.pocket_finding.dogsitescorer import find_pocket_dogsitescorer
 from scripts.pocket_finding.manual import parse_pocket_coordinates
 from scripts.pocket_finding.p2rank import find_pocket_p2rank
 from scripts.pocket_finding.radius_of_gyration import find_pocket_RoG
-from scripts.utilities import printlog
+from scripts.utilities.utilities import printlog
 
+POCKET_DETECTION_OPTIONS = ['Reference', 'RoG', 'Dogsitescorer', 'p2rank', 'Manual']
 
-def pocket_finder(mode: str, software: Path = None, receptor: Path = None, ligand: Path = None, radius: int = 10):
+def pocket_finder(mode: str, software: Path = None, receptor: Path = None, ligand: Path = None, radius: int = 10, manual_pocket: str = None):
     # Determine the docking pocket
     if mode == 'Reference':
         pocket_definition = find_pocket_default(ligand, receptor, radius)
@@ -24,7 +25,7 @@ def pocket_finder(mode: str, software: Path = None, receptor: Path = None, ligan
         pocket_definition = find_pocket_dogsitescorer(receptor, method='volume')
     elif mode == 'p2rank':
         pocket_definition = find_pocket_p2rank(software, receptor, radius)
-    else:
+    elif mode == 'Manual':
         pocket_definition = parse_pocket_coordinates(mode)
     printlog(f'Pocket definition: {pocket_definition}')
     return pocket_definition
