+++ scripts/pocket_finding/p2rank.py
@@ -12,7 +12,7 @@ import requests
 dockm8_path = next((p / 'DockM8' for p in Path(__file__).resolve().parents if (p / 'DockM8').is_dir()), None)
 sys.path.append(str(dockm8_path))
 
-from scripts.utilities import printlog
+from scripts.utilities.utilities import printlog
 
 
 def find_pocket_p2rank(software: Path, receptor: Path, radius: int):
