+++ scripts/pocket_finding/radius_of_gyration.py
@@ -13,7 +13,7 @@ from scripts.pocket_finding.utils import (
     get_ligand_coordinates,
     process_protein_and_ligand,
 )
-from scripts.utilities import load_molecule, printlog
+from scripts.utilities.utilities import load_molecule, printlog
 
 
 def find_pocket_RoG(ligand_file: Path, protein_file: Path):
