+++ scripts/protein_preparation/fetching/fetch_alphafold.py
@@ -6,7 +6,7 @@ import sys
 dockm8_path = next((p / 'DockM8' for p in Path(__file__).resolve().parents if (p / 'DockM8').is_dir()), None)
 sys.path.append(str(dockm8_path))
 
-from scripts.utilities import printlog
+from scripts.utilities.utilities import printlog
 
 def fetch_alphafold_structure(uniprot_code: str, output_dir: Path):
     """
