+++ scripts/protein_preparation/main.py
@@ -17,12 +17,11 @@ from scripts.protein_preparation.protonation.protonate_protoss import (
     protonate_protein_protoss,
 )
 from scripts.protein_preparation.structure_assessment.edia import get_best_chain_edia
-from scripts.utilities import printlog
+from scripts.utilities.utilities import printlog
 
 
 def prepare_protein(
     protein_file_or_code: Path,
-    type: str = "File",
     output_dir: Path = None,
     select_best_chain: bool = True,
     fix_protein: bool = True,
@@ -38,10 +37,8 @@ def prepare_protein(
 
     Args:
         protein_file_or_code (str or Path): The protein_file_or_code value. It can be a PDB code, Uniprot code, or file path.
-        type (str, optional): The type of protein_file_or_code. Can be 'PDB', 'Uniprot', or 'File'. Default is 'File'.
         output_dir (str or Path, optional): The directory where the prepared protein structure will be saved. If not provided, the same directory as the protein_file_or_code file will be used.
         select_best_chain (bool, optional): Whether to select the best chain from the protein_file_or_code structure. Only applicable for PDB protein_file_or_code. Default is True.
-        fix_protein (bool, optional): Whether to fix the protein structure. Default is True.
         fix_nonstandard_residues (bool, optional): Whether to fix nonstandard residues in the protein structure. Default is True.
         fix_missing_residues (bool, optional): Whether to fix missing residues in the protein structure. Default is True.
         add_missing_hydrogens_pH (float, optional): The pH value for adding missing hydrogens. Default is 7.0.
@@ -53,17 +50,16 @@ def prepare_protein(
         Path: The path to the prepared protein structure.
     """
 
-    if not type:
-        if len(protein_file_or_code) == 4 and protein_file_or_code.isalnum():
-            type = "PDB"
-        elif len(protein_file_or_code) == 6 and protein_file_or_code.isalnum():
-            type = "Uniprot"
+    if len(protein_file_or_code) == 4 and protein_file_or_code.isalnum():
+        type = "PDB"
+    elif len(protein_file_or_code) == 6 and protein_file_or_code.isalnum():
+        type = "Uniprot"
+    else:
+        # Check if the protein_file_or_code is a valid path
+        if not Path(protein_file_or_code).is_file():
+            raise ValueError("Protein_file_or_code file is an invalid file path.")
         else:
-            # Check if the protein_file_or_code is a valid path
-            if not Path(protein_file_or_code).is_file():
-                raise ValueError("Protein_file_or_code file is an invalid file path.")
-            else:
-                type = "File"
+            type = "File"
 
     output_dir.mkdir(parents=True, exist_ok=True)
 
@@ -107,9 +103,7 @@ def prepare_protein(
         step1_pdb = Path(protein_file_or_code)
 
     # Fix the protein structure
-    if (
-        fix_protein
-        or fix_nonstandard_residues
+    if (fix_nonstandard_residues
         or fix_missing_residues
         or add_missing_hydrogens_pH is not None
         or remove_hetero
